requirements.txt
Flask==3.1.2
pymongo==4.16.0
dnspython==2.8.0
python-dotenv==1.2.2
scikit-learn==1.8.0
numpy==2.4.1
scipy==1.17.0
pdfminer.six==20260107
python-docx==1.2.0
certifi
gunicorn